Move PrecomputeAddress to contract package
The PrecomputeAddress function precompute the address of a contract
instance thus it's not directly related to the 'account' package,
but to the 'contract' package. The only use of this function to the
'account' package is to calculate the account contract address for
deploying an account, and to satisfy that I created the
PrecomputeAccountAddress function in the 'account' package that
calls the PrecomputeAddress function passing the proper parameters.